allow ts segments with zero frames
the number of ts segments is derived from the track durations in the mp4
atoms. it is possible that the duration is for example 10 segments + 1/2
frame. in this case, the m3u8 will point to 11 segments, but the last
segment will have no frames (the last frame is already included in segment
number 10)