concat clip - allow offset larger than range end
in live recording when the video stream starts after the audio, it is
possible that the end offset of the segment will be snapped to the video
start offset. in this case, end will be equal to offset.
when this happens, just return an empty clip (range.start==range.end)