cache fetch optimization
no need to copy the buffer when reading from drm/mapping cache.
when parsing a json, the strings values/keys keep a pointer to the
original buffer, but in all cases when the string is retained, it is
copied.
also, no need to save the mapping cache with a null terminator since that
is already handled by the buffer cache