live -> vod - clipTo bug fix
when the original playlist type is live, both clipFrom and clipTo use
the live timestamps. therefore, need to use original_times to calculate
last_clip_end, even when the final playlist type is vod.
for a playlist that was originally vod, the value of last_clip_end will
remain exactly as it was.