VOD-5061: caption position (#68)
We are over specifying parameters for captions and it's causing iOS and safari to display captions in the middle of the screen. I am removing some parameters and changing the position of the captions so are near the bottom of the screen.

Reviewers: @cobarx @dchamberlin-cr

This PR needs will be followed by a PR to update evs-nginx-vod to this new commit.